\name{primorial}
\alias{primorial}
\title{Primorial}
\description{
Primorial - The Factorial of Prime Numbers.
}
\usage{
factorial(n)
}
\arguments{
  \item{n}{A non-negative integer}
}
\details{
  For the \eqn{n}th prime number \eqn{p_n# = p_1 * p_2*...*p_k} with \eqn{k < n}

}
\examples{
primorial(12) = 1 * 2 * 3 * 5 * 7 * 11 = 2310
}
