\name{tau}
\alias{tau}
\title{
  number of all divisors of an integer n.
  }
\description{
  τ(n): Number of all divisors of a natural number n.
  }
\usage{
  tau(n)
  }
\arguments{
  \item{n}{A non-negative integer}
  }
\details{
  Works well up to around 10⁹.
  }
\examples{
  tau(100) = 9
  tau(10)  = 4   # Divisors of 10 are 1, 2, 5,10)
  }
