\name{sigma}
\alias{sigma}
\title{
  Sum of all divisors of a natural number n
  }
\description{
  Sum of all divisors of a natural number n.
  }
\usage{
  sigma(n)
  }
\arguments{
  \item{n}{A non-negative integer}
  }
\details{
  Works well up to around 10⁹.
  }
\examples{
  sigma(100)=217
  sigma(10) =118 (=1+2+5+10)
  }
